:truck: rename: member용 응답 member 폴더로 분리
