+++ src/main/java/umc/product/domain/notice/controller/NoticeController.java
@@ -10,8 +10,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import umc.product.domain.member.entity.Member;
 import umc.product.domain.notice.adviser.NoticeAdviser;
-import umc.product.domain.notice.dto.response.member.list.NoticeCheckResponse;
-import umc.product.domain.notice.dto.response.member.list.NoticeDetailResponse;
+import umc.product.domain.notice.dto.response.member.NoticeCheckResponse;
+import umc.product.domain.notice.dto.response.member.NoticeDetailResponse;
 import umc.product.domain.notice.dto.response.member.list.NoticeListResponse;
 import umc.product.domain.notice.entity.enums.NoticeTarget;
 import umc.product.global.common.base.BaseResponse;
