+++ src/main/java/umc/product/domain/notice/dto/response/admin/list/AdminNoticeResponse.java
@@ -1,0 +1,4 @@
+package umc.product.domain.notice.dto.response.admin.list;
+
+public record AdminNoticeResponse() {
+}
