+++ src/main/java/umc/product/domain/notice/dto/response/member/NoticeResponse.java
@@ -1,4 +1,4 @@
-package umc.product.domain.notice.dto.response.member.list;
+package umc.product.domain.notice.dto.response.member;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
