+++ src/main/java/umc/product/domain/notice/dto/response/member/list/NoticeListResponse.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import umc.product.domain.notice.dto.response.member.NoticeResponse;
 
 import java.util.List;
